refactor: Reorganize core application files and templates into `app/core` and `app/common` directories.
