+++ app/core/templates/setup.html
