+++ app/modes/chat/routes.py
@@ -1,7 +1,7 @@
 from flask import render_template, request, redirect, url_for
 from . import chat_bp
-from app.core.storage import load_topic, save_chat_history, save_topic
-from app.core.agents import PlannerAgent
+from app.common.storage import load_topic, save_chat_history, save_topic
+from app.common.agents import PlannerAgent
 from app.modes.chat.agent import ChatModeMainChatAgent
 from app.modes.chapter.agent import ChapterModeChatAgent
 
@@ -21,7 +21,7 @@ def mode(topic_name):
 
     if not chat_history:
         # Generate welcome message if chat is new
-        from app.core.utils import get_user_context
+        from app.common.utils import get_user_context
         user_background = get_user_context()
         
         # 1. Generate Plan if missing
@@ -68,14 +68,14 @@ def update_plan(topic_name):
         return redirect(url_for('chat.mode', topic_name=topic_name))
 
     current_plan = topic_data.get('plan', [])
-    from app.core.utils import get_user_context
+    from app.common.utils import get_user_context
     user_background = get_user_context()
 
     planner = PlannerAgent()
     # Call agent to get a new plan
     new_plan, error = planner.update_study_plan(topic_name, user_background, current_plan, comment)
 
-    from app.core.utils import reconcile_plan_steps
+    from app.common.utils import reconcile_plan_steps
 
     if not error:
         # Save the new plan
@@ -118,7 +118,7 @@ def send_message(topic_name):
         context = f'The topic is {topic_name}. No additional details are available yet.'
         plan = []
     
-    from app.core.utils import get_user_context
+    from app.common.utils import get_user_context
     user_background = get_user_context()
 
     # Load history from DB
@@ -162,7 +162,7 @@ def chat(topic_name, step_index):
     step_history = current_step_data.get('chat_history', [])
     step_history.append({"role": "user", "content": user_question})
     
-    from app.core.utils import get_user_context
+    from app.common.utils import get_user_context
     current_background = get_user_context()
     
     # Pass the history to the agent.
