+++ app/modes/quiz/agent.py
@@ -1,6 +1,6 @@
 import json
 from datetime import datetime
-from app.core.utils import call_llm, validate_quiz_structure
+from app.common.utils import call_llm, validate_quiz_structure
 
 class QuizAgent:
     """
@@ -88,7 +88,7 @@ class QuizAgent:
         Returns (count, None) on success or (default_count, error) on failure.
         """
         if user_background is None:
-            from app.core.utils import get_user_context
+            from app.common.utils import get_user_context
             user_background = get_user_context()
 
         from app.modes.quiz.prompts import get_quiz_count_prompt
