+++ migrate_data.py
@@ -6,7 +6,7 @@ from dotenv import load_dotenv
 load_dotenv()
 
 from app import create_app, db  # noqa: E402
-from app.core import storage  # noqa: E402
+from app.common import storage  # noqa: E402
 
 def migrate():
     """
